requirements.txt
pyTelegramBotAPI
transformers
accelerate
bitsandbytes
langchain
sentence-transformers
faiss-gpu
python-dotenv
kaggle
pandas
numpy